get_adjacency_mat {txmBioinfoToolkit}	R Documentation

Get adjacency matrix from a sample-feature matrix

Description

Get adjacency matrix from a sample-feature matrix

Usage

get_adjacency_mat(data, k = 20)

Arguments

data
a matrix whose rows represent samples and columns represent features

k
how many k nearest neighbors to build adjacency matrix

Value

Returns a adjacency matrix

[Package txmBioinfoToolkit version 0.1.0 Index]