my_RF {txmBioinfoToolkit}	R Documentation

Random forest code memo

Description

Random forest code memo

Usage

my_RF(train, labels, test, verbose = T)

Arguments

train
The train dataset whose rows are samples and columns are features

labels
Labels of the train dataset

test
The test dataset whose rows are samples and columns are features

verbose
Whether to display messages

Value

todo

[Package txmBioinfoToolkit version 0.1.0 Index]