my_knn {txmBioinfoToolkit}	R Documentation

Several K Nearest Neighbors (KNN) methods

Description

Several K Nearest Neighbors (KNN) methods

Usage

my_knn(
  train,
  labels,
  test,
  k,
  method = "weighted",
  dist.metric = "cosine",
  verbose = T
)

Arguments

train
The train dataset whose rows are samples and columns are features

labels
Labels of the train dataset

test
The test dataset whose rows are samples and columns are features

k
Number of neighbors

method
Knn method; can be one of "weighted", "Shieh", "normal"

dist.metric
Distance metric; can be one of "euclidean", "cosine", "manhattan",

verbose
Whether to display messages

Value

Return a data frame containing a predicted labels column and a score column.

[Package txmBioinfoToolkit version 0.1.0 Index]