mykmeans {txmBioinfoToolkit}	R Documentation

Perform kmeans with multiple 'k's and automatically select the best k

Description

Perform kmeans with multiple 'k's and automatically select the best k with the maximum average silhouette

Usage

mykmeans(x, k, iter.max = 10)

Arguments

x
numeric matrix of data, or an object that can be coerced to such a matrix.

k
a set of candidate centers.

iter.max
the maximum number of iterations allowed.

Value

returns an object of class "kmeans".

[Package txmBioinfoToolkit version 0.1.0 Index]